multimodal-rag-embed-2/pyproject.toml
[project]
name = "multimodal-rag-embed-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.7.0",
    "google-genai-haystack>=3.11.0",
    "haystack-ai>=2.25.2",
    "pillow>=12.1.1",
    "pypdf>=6.8.0",
    "pypdfium2>=5.6.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
]
